Bang out some more tests.

+++ src/Nines/BlogBundle/Controller/PageController.php
@@ -28,8 +28,9 @@ class PageController extends Controller
     public function indexAction(Request $request)
     {
         $em = $this->getDoctrine()->getManager();
-        $dql = 'SELECT e FROM NinesBlogBundle:Page e ORDER BY e.id';
-        $query = $em->createQuery($dql);
+        $private = $this->get('security.authorization_checker')->isGranted('ROLE_ADMIN');
+        $repo = $em->getRepository(Page::class);
+        $query = $repo->listQuery($private);
         $paginator = $this->get('knp_paginator');
         $pages = $paginator->paginate($query, $request->query->getint('page', 1), 25);
 
@@ -109,6 +110,10 @@ class PageController extends Controller
      */
     public function newAction(Request $request)
     {
+        if( ! $this->isGranted('ROLE_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $user = $this->getUser();
         $page = new Page();
         $page->setUser($user);
@@ -146,6 +151,9 @@ class PageController extends Controller
      */
     public function showAction(Page $page)
     {
+        if( ! $page->getPublic()) {
+            $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'Unable to access this page!');            
+        }
 
         return array(
             'page' => $page,
@@ -163,6 +171,7 @@ class PageController extends Controller
      */
     public function editAction(Request $request, Page $page)
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'Unable to access this page!');            
         $editForm = $this->createForm('Nines\BlogBundle\Form\PageType', $page);
         $editForm->handleRequest($request);
 
@@ -194,6 +203,7 @@ class PageController extends Controller
      */
     public function deleteAction(Request $request, Page $page)
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'Unable to access this page!');            
         $em = $this->getDoctrine()->getManager();
         $em->remove($page);
         $em->flush();
